@@ -1,5 +1,4 @@
-//go:build go1.17
-// +build go1.17
+//go:build go1.18
 
 // Copyright 2014 The Gogs Authors. All rights reserved.
 // Use of this source code is governed by a MIT-style
